refactored server query code into Utils.js
