+++ chrome-extension/html/background.html
@@ -7,5 +7,6 @@
 <body>
   <!-- <script type="text/javascript" src="../js/eventRegisterTests.js"></script> -->
   <script type="module" src="../js/BackgroundScript.js"></script>
+  <!-- <script type="module" src="../js/Utit.js"></script> -->
 </body>
 </html>
